fix: review UI, closing bounty, wallet check

+++ src/server/routes/bounties/deleteBounty.ts
@@ -1,9 +1,11 @@
-import { BountyState } from '@prisma/client';
+import { BountyState, SubmissionState } from '@prisma/client';
 import { TRPCError } from '@trpc/server';
 import { z } from 'zod';
 
 import { LOG } from '@/server/importer/utils';
 import { staffOnlyProcedure } from '@/server/procedures';
+
+import { CloseBounty } from './updateBounty';
 
 const DeleteBountySchema = z.object({
   slug: z.string(),
@@ -18,6 +20,9 @@ export const deleteBounty = staffOnlyProcedure
       where: {
         slug: input.slug,
       },
+      include: {
+        submissions: true,
+      },
     });
     if (!bounty) {
       throw new TRPCError({
@@ -31,10 +36,24 @@ export const deleteBounty = staffOnlyProcedure
         message: 'Cannot delete closed bounty',
       });
     }
-    const deletedBounty = await prisma.bounty.delete({
-      where: {
-        slug: input.slug,
-      },
-    });
-    LOG(`${bounty?.slug}: Deleted bounty "${bounty?.title}"`);
+    if (bounty.submissions.length > 0) {
+      if (
+        bounty.submissions.filter(
+          (sub) => sub.state === SubmissionState.WaitingForReview
+        ).length > 0
+      ) {
+        throw new TRPCError({
+          code: 'FORBIDDEN',
+          message: 'Cannot delete bounty with open submissions',
+        });
+      }
+      CloseBounty(prisma, input.slug);
+    } else {
+      const deletedBounty = await prisma.bounty.delete({
+        where: {
+          slug: input.slug,
+        },
+      });
+      LOG(`${bounty?.slug}: Deleted bounty "${bounty?.title}"`);
+    }
   });
